feat: implement identity v2

+++ src/spaceone/identity/service/provider_service.py
@@ -3,6 +3,7 @@ from typing import Union
 
 from spaceone.core.service import *
 from spaceone.core.service.utils import *
+from spaceone.core.error import *
 
 from spaceone.identity.model.provider.request import *
 from spaceone.identity.model.provider.response import *
@@ -11,20 +12,21 @@ from spaceone.identity.manager.provider_manager import ProviderManager
 _LOGGER = logging.getLogger(__name__)
 
 
+@authentication_handler
+@authorization_handler
+@mutation_handler
+@event_handler
 class ProviderService(BaseService):
-
-    service = "identity"
     resource = "Provider"
-    permission_group = "DOMAIN"
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.provider_mgr = ProviderManager()
 
-    @transaction(scope="domain_admin:write")
+    @transaction(permission="identity:Provider.write", role_types=["DOMAIN_ADMIN"])
     @convert_model
     def create(self, params: ProviderCreateRequest) -> Union[ProviderResponse, dict]:
-        """ create provider
+        """create provider
 
         Args:
             params (ProviderCreateRequest): {
@@ -36,7 +38,7 @@ class ProviderService(BaseService):
                 'order': 'int',
                 'options': 'dict',
                 'tags': 'dict',
-                'domain_id': 'str'      # required
+                'domain_id': 'str'      # injected from auth
             }
 
         Returns:
@@ -46,10 +48,10 @@ class ProviderService(BaseService):
         provider_vo = self.provider_mgr.create_provider(params.dict())
         return ProviderResponse(**provider_vo.to_dict())
 
-    @transaction(scope="domain_admin:write")
+    @transaction(permission="identity:Provider.write", role_types=["DOMAIN_ADMIN"])
     @convert_model
     def update(self, params: ProviderUpdateRequest) -> Union[ProviderResponse, dict]:
-        """ update provider
+        """update provider
 
         Args:
             params (ProviderUpdateRequest): {
@@ -61,7 +63,7 @@ class ProviderService(BaseService):
                 'order': 'int',
                 'options': 'dict',
                 'tags': 'dict',
-                'domain_id': 'str'      # required
+                'domain_id': 'str'      # injected from auth
             }
 
         Returns:
@@ -69,6 +71,8 @@ class ProviderService(BaseService):
         """
 
         provider_vo = self.provider_mgr.get_provider(params.provider, params.domain_id)
+        if provider_vo.is_managed:
+            raise ERROR_PERMISSION_DENIED()
 
         provider_vo = self.provider_mgr.update_provider_by_vo(
             params.dict(exclude_unset=True), provider_vo
@@ -76,15 +80,15 @@ class ProviderService(BaseService):
 
         return ProviderResponse(**provider_vo.to_dict())
 
-    @transaction(scope="domain_admin:write")
+    @transaction(permission="identity:Provider.write", role_types=["DOMAIN_ADMIN"])
     @convert_model
     def delete(self, params: ProviderDeleteRequest) -> None:
-        """ delete provider
+        """delete provider
 
         Args:
             params (ProviderDeleteRequest): {
                 'provider': 'str',      # required
-                'domain_id': 'str'      # required
+                'domain_id': 'str'      # injected from auth
             }
 
         Returns:
@@ -92,17 +96,23 @@ class ProviderService(BaseService):
         """
 
         provider_vo = self.provider_mgr.get_provider(params.provider, params.domain_id)
+        if provider_vo.is_managed:
+            raise ERROR_PERMISSION_DENIED()
+
         self.provider_mgr.delete_provider_by_vo(provider_vo)
 
-    @transaction(scope="workspace_member:read")
+    @transaction(
+        permission="identity:Provider.read",
+        role_types=["DOMAIN_ADMIN", "WORKSPACE_OWNER", "WORKSPACE_MEMBER"],
+    )
     @convert_model
     def get(self, params: ProviderGetRequest) -> Union[ProviderResponse, dict]:
-        """ delete provider
+        """delete provider
 
         Args:
             params (ProviderGetRequest): {
                 'provider': 'str',      # required
-                'domain_id': 'str'      # required
+                'domain_id': 'str'      # injected from auth
             }
 
         Returns:
@@ -112,12 +122,17 @@ class ProviderService(BaseService):
         provider_vo = self.provider_mgr.get_provider(params.provider, params.domain_id)
         return ProviderResponse(**provider_vo.to_dict())
 
-    @transaction(scope="workspace_member:read")
-    @append_query_filter(['provider', 'name', 'alias', 'is_managed', 'domain_id'])
-    @append_keyword_filter(['provider', 'name'])
+    @transaction(
+        permission="identity:Provider.read",
+        role_types=["DOMAIN_ADMIN", "WORKSPACE_OWNER", "WORKSPACE_MEMBER"],
+    )
+    @append_query_filter(["provider", "name", "alias", "is_managed", "domain_id"])
+    @append_keyword_filter(["provider", "name"])
     @convert_model
-    def list(self, params: ProviderSearchQueryRequest) -> Union[ProvidersResponse, dict]:
-        """ list providers
+    def list(
+        self, params: ProviderSearchQueryRequest
+    ) -> Union[ProvidersResponse, dict]:
+        """list providers
 
         Args:
             params (ProviderSearchQueryRequest): {
@@ -126,7 +141,7 @@ class ProviderService(BaseService):
                 'name': 'str',
                 'alias': 'str',
                 'is_managed': 'bool',
-                'domain_id': 'str'      # required
+                'domain_id': 'str'      # injected from auth
             }
 
         Returns:
@@ -134,22 +149,27 @@ class ProviderService(BaseService):
         """
 
         query = params.query or {}
-        provider_vos, total_count = self.provider_mgr.list_providers(query, params.domain_id)
+        provider_vos, total_count = self.provider_mgr.list_providers(
+            query, params.domain_id
+        )
 
         providers_info = [provider_vo.to_dict() for provider_vo in provider_vos]
         return ProvidersResponse(results=providers_info, total_count=total_count)
 
-    @transaction(scope="workspace_member:read")
-    @append_query_filter(['domain_id'])
-    @append_keyword_filter(['provider', 'name'])
+    @transaction(
+        permission="identity:Provider.read",
+        role_types=["DOMAIN_ADMIN", "WORKSPACE_OWNER", "WORKSPACE_MEMBER"],
+    )
+    @append_query_filter(["domain_id"])
+    @append_keyword_filter(["provider", "name"])
     @convert_model
     def stat(self, params: ProviderStatQueryRequest) -> dict:
-        """ stat providers
+        """stat providers
 
         Args:
             params (ProviderStatQueryRequest): {
                 'query': 'dict (spaceone.api.core.v1.StatisticsQuery)', # required
-                'domain_id': 'str'    # required
+                'domain_id': 'str'    # injected from auth
             }
 
         Returns:
